Simplify access to library by introducing Core struct. Calling NewCore function opens the given device and obtains the Identity and Level0Discovery struct from the given device.
